pyproject.toml
[tool.poetry]
name = "tilelite"
version = "0.1.0-rc1"
description = "tiledesk lite"
authors = ["Gianluca Lorenzo <gianluca.lorenzo@gmail.com>"]
repository = "https://github.com/Tiledesk/tiledesk-llm"
license = "MIT"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.115.11" #"0.114.2" #"0.111.1"
jsonschema= "4.23.0"
jq = "1.8.0"

langchain =  "0.3.20" # "0.3.19"#"0.3.17"#"0.3.10" #"0.3.1"#"0.3.0" #"0.2.16"

openai = "1.66.3" #"1.64.0"
langchain-openai = "0.3.8" #"0.3.7"#"0.3.3"#"0.2.1" #"0.2.0" #"0.1.19"
langchain-voyageai = "0.1.3"#"0.1.2" #"0.1.1"
langchain-anthropic = "0.3.9"#"0.3.8"
langchain-cohere= "0.4.2"#"0.3.0" #"0.1.9"
langchain-google-genai= "2.1.0" #"2.0.10" #"2.0.11"#"2.0.9"#"2.0.0" #"1.0.8"
langchain-groq = "0.2.5" #"0.2.4"#"0.2.0" #"0.1.8"
langchain-aws= "0.2.15"#"0.2.1" #"0.1.12"
langchain-deepseek = "0.1.2"
langchain-community = "0.3.19"#"0.3.18"#"0.3.16"#"0.3.1" #"0.2.10"
langchain-experimental = "0.3.4"# "0.3.1" #no previous
langchain-ollama= "0.2.3"#"0.2.0"
#peft = "0.14.0"#"0.13.0"

tiktoken = "0.9.0"#"0.7.0"

#uvicorn = "^0.28"
psutil= "6.0.0"
httpx= "0.27.2" #"0.27.0"
gql= "3.5.2"#"3.5.0"
PyJWT= "2.10.1"#"2.9.0"
#torch="2.4.1"


[tool.poetry.dependencies.uvicorn]
version = "0.34.0" #"0.30.3"
extras = ["standard"]
# Add other dependencies here as needed
#[tool.poetry.dependencies.unstructured]
#version = "^0.12.6"
#extras = [ "pdf"]

[tool.poetry.group.dev.dependencies.uvicorn]
version = "0.34.0" #"0.30.3"
extras = ["standard"]

[tool.poetry.scripts]
tilelite = "tilelite.__main__:main"
